Move dialog-related code to a single file
