+++ pynicotine/gtkgui/dialogs.py
@@ -2,6 +2,7 @@
 # COPYRIGHT (C) 2016-2017 Michael Labouebe <gfarmerfr@free.fr>
 # COPYRIGHT (C) 2008-2009 Quinox <quinox@users.sf.net>
 # COPYRIGHT (C) 2006-2009 Daelstorm <daelstorm@gmail.com>
+# COPYRIGHT (C) 2003-2004 Hyriand <hyriand@thegraveyard.org>
 #
 # GNU GENERAL PUBLIC LICENSE
 #    Version 3, 29 June 2007
@@ -19,8 +20,15 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import os
+
+from gettext import gettext as _
+
 from gi.repository import GObject
 from gi.repository import Gtk
+
+
+""" General Dialogs """
 
 
 def combo_box_dialog(parent, title, message, default_text="",
@@ -103,6 +111,22 @@ def entry_dialog(parent, title, message, default=""):
     return result
 
 
+def message_dialog(parent, title, message):
+
+    self = Gtk.MessageDialog(
+        transient_for=parent,
+        flags=0,
+        message_type=Gtk.MessageType.INFO,
+        buttons=Gtk.ButtonsType.OK,
+        text=title
+    )
+    self.set_modal(True)
+    self.format_secondary_text(message)
+
+    self.run()
+    self.destroy()
+
+
 def option_dialog(parent, title, message, callback, callback_data=None, checkbox_label="", third=""):
 
     self = Gtk.MessageDialog(
@@ -126,3 +150,123 @@ def option_dialog(parent, title, message, callback, callback_data=None, checkbox
         self.add_button(third, Gtk.ResponseType.REJECT)
 
     self.show()
+
+
+""" File Chooser Dialogs """
+
+
+def choose_dir(parent=None, initialdir="~", title=None, multichoice=True):
+    try:
+        dialog = Gtk.FileChooserNative.new(
+            title,
+            parent,
+            Gtk.FileChooserAction.SELECT_FOLDER,
+            _("_Open"),
+            _("_Cancel")
+        )
+    except AttributeError:
+        dialog = Gtk.FileChooserDialog(
+            title,
+            parent,
+            Gtk.FileChooserAction.SELECT_FOLDER
+        )
+        dialog.add_buttons(_("_Cancel"), Gtk.ResponseType.CANCEL, _("_Open"), Gtk.ResponseType.ACCEPT)
+
+    if multichoice:
+        dialog.set_select_multiple(True)
+
+    folder = os.path.expanduser(initialdir)
+
+    if os.path.exists(folder):
+        dialog.set_current_folder(folder)
+    else:
+        dialog.set_current_folder(os.path.expanduser("~"))
+
+    response = dialog.run()
+
+    if response == Gtk.ResponseType.ACCEPT:
+        res = dialog.get_filenames()
+    else:
+        res = None
+
+    dialog.destroy()
+
+    return res
+
+
+def choose_file(parent=None, initialdir="~", initialfile="", multiple=False):
+    try:
+        dialog = Gtk.FileChooserNative.new(
+            None,
+            parent,
+            Gtk.FileChooserAction.OPEN,
+            _("_Open"),
+            _("_Cancel")
+        )
+    except AttributeError:
+        dialog = Gtk.FileChooserDialog(
+            None,
+            parent,
+            Gtk.FileChooserAction.OPEN
+        )
+        dialog.add_buttons(_("_Cancel"), Gtk.ResponseType.CANCEL, _("_Open"), Gtk.ResponseType.ACCEPT)
+
+    dialog.set_select_multiple(multiple)
+    folder = os.path.expanduser(initialdir)
+
+    if os.path.exists(folder):
+        dialog.set_current_folder(folder)
+    else:
+        dialog.set_current_folder(os.path.expanduser("~"))
+
+    response = dialog.run()
+
+    if response == Gtk.ResponseType.ACCEPT:
+        res = dialog.get_filenames()
+    else:
+        res = None
+
+    dialog.destroy()
+
+    return res
+
+
+def save_file(parent=None, initialdir="~", initialfile="", title=None):
+    try:
+        dialog = Gtk.FileChooserNative.new(
+            title,
+            parent,
+            Gtk.FileChooserAction.SAVE,
+            _("_Save"),
+            _("_Cancel")
+        )
+    except AttributeError:
+        dialog = Gtk.FileChooserDialog(
+            title,
+            parent,
+            Gtk.FileChooserAction.SAVE
+        )
+        dialog.add_buttons(_("_Cancel"), Gtk.ResponseType.CANCEL, _("_Save"), Gtk.ResponseType.ACCEPT)
+
+    dialog.set_select_multiple(False)
+    dialog.set_show_hidden(True)
+
+    folder = os.path.expanduser(initialdir)
+
+    if os.path.exists(folder):
+        dialog.set_current_folder(folder)
+    else:
+        dialog.set_current_folder(os.path.expanduser("~"))
+
+    dialog.set_current_name(initialfile)
+
+    response = dialog.run()
+
+    if response == Gtk.ResponseType.ACCEPT:
+        res = dialog.get_filenames()
+    else:
+        res = None
+
+    dialog.destroy()
+
+    return res
