+++ pynicotine/gtkgui/fastconfigure.py
@@ -28,7 +28,7 @@ from gi.repository import GObject
 from gi.repository import Gtk
 
 import _thread
-from pynicotine.gtkgui.dirchooser import choose_dir
+from pynicotine.gtkgui.dialogs import choose_dir
 from pynicotine.gtkgui.dialogs import combo_box_dialog
 from pynicotine.gtkgui.utils import human_size
 from pynicotine.gtkgui.utils import initialise_columns
